Share the shared stuff
