+++ app/src/lib/shells/index.ts
@@ -1,7 +1,1 @@
-import * as Darwin from './darwin'
-import * as Win32 from './win32'
-
-export type Shell = Darwin.Shell | Win32.Shell
-
-export { getAvailableShells } from './lookup'
-export { launchShell } from './launch'
+export * from './shared'
