+++ app/src/lib/shells/shared.ts
@@ -1,6 +1,27 @@
 import * as Darwin from './darwin'
 import * as Win32 from './win32'
-import { Shell } from './index'
+
+export type Shell = Darwin.Shell | Win32.Shell
+
+let shellCache: ReadonlyArray<Shell> | null = null
+
+export async function getAvailableShells(): Promise<ReadonlyArray<Shell>> {
+  if (shellCache) {
+    return shellCache
+  }
+
+  if (__DARWIN__) {
+    shellCache = await Darwin.getAvailableShells()
+    return shellCache
+  } else if (__WIN32__) {
+    shellCache = await Win32.getAvailableShells()
+    return shellCache
+  }
+
+  return Promise.reject(
+    `Platform not currently supported for resolving shells: ${process.platform}`
+  )
+}
 
 export async function launchShell(shell: Shell) {
   // We have to manually cast the wider `Shell` type into the platform-specific
